requirements.txt
python-telegram-bot>=20.8,<21.0
httpx>=0.26.0,<0.27.0
websockets>=12.0
python-dotenv>=1.0.1
feedparser>=6.0.11
yfinance>=0.2.40
openai>=1.0.0
